HUB-237: Use absolute position for 2nd level links
Before this change, the second level links were fixed and had only left
border. This looked little bit broken when scrolling the page.

This change will use aboslute position for the second level leaving the
links to the top position and allowing user to scroll the page without
floating second level menu. Also the borders in bottom and right makes
it look more contained in the menu.