HUB-262: Use submit event instead click event
In Firefox the click event did not cause the search form to be
submitted. Our guess is that the previous trigger event is
non-blocking call which therefore is calling click event too early
and since the submit button is disabled, it wont submit the form.

This change should call form submission despite the disabled state of
submit button.