HUB-384: Restore patches directory
Infrastructure configurations are currently assuming this directory exists and
is doing symlinkin as part of the build. If we remove this directory we get an
error saying that directory does not exist.

We could remove this directory symlinking from infra configs, but we decided to
have this plaeholcer at least for a while until this branch lands on other
branches too (therefore avoiding errors where one can complain about missing
patch).