HUB-423: Fix menu layout on course embed pages
The course embed React component imports the full styleguide. This
overrides (reverts) some of the styles that the Guide has already
overridden causing layout breaking on secondary menu. Make some of the
Guide styles important to enforce them. Later on we should look into
removing most of the Guide's overrides and use as much of the styleguide
styles as possible and only do minimal overrides (if this is feasible).