HUB-427: Fix filtering by degree programme and other education provider
There is a hard-coded entity type check that assumes that index zero
exists when looping items. This assumption fails when the index starts
from non-zero number, such as 1. The index can start from non-zero
number when filtering has removed, for example, the first item. Thus the
filtering happening after that will not have an item in the unset index
(as the filtering does not reset array indices). Removing the check
seems to have no undesired side effects and seems to achieve the desired
filtering.