HUB-725 - Documented the new domains:
https://guide.student.helsinki.fi/ (old)
https://studies.helsinki.fi/ (new)

Changes in may 2020
There is a new domain set for student guide: studies-qa.it.helsinki.fi .
Guide will be sharing the studies domain which also hosts multiple other services
in the same root. Drupal is accessed via a proxy which reserves Drupal related
language prefixed paths to student guide. Teaching instructions still remain
in the old domain, which further complicates the setup.

The new domains are:
  - local.studies-qa.it.helsinki.fi
  - dev.studies-qa.it.helsinki.fi
  - studies.helsinki.fi

NOTE: most of the changes are in the separate student_guide_infra repo.