HUB-755 - Onelogin/saml with old samlauth module is quite probablematic in the project.
For example updating with composer update may be very difficult locally depending on PHP-version,
as the old samlauth depends on mcrypt, which is deprecated. Samlauth may stop any composer command
unless its run with:
composer require onelogin/php-saml:v2.13.0 --ignore-platform-reqs

The composer.lock changes are a result of this.

We hopefully can refactor the login process soon, either dropping samlauth and many customizations
or going for an upgrade (similar to courses 2.0 auth).