HUB-755 - Oprek mock responses now have 2 choices, simple and complex. There are real world but anonymized use cases and the complex one includes the exact bug we trying to fix here.
Its clear that the bug is not about finding the study right and degree programme, as those are actually set correctly for the person affected.
What is not set correctly is the "is_primary" field in the personal flagging of the degree programme.

"Ensisijainen" / "Optio" are read correctly but it appears in some cases the primary flag may be overridden or merged mistakenly in getTargetedCodes().