HUB-755 - UserSync -> study right/code check debugging now shows less but more relevant data. This was actually quite useful in getting to the bottom of the problem.
The issue is simply that a primary study right is actually found always, and its either a combination
of parent and child (preferred) or just the child. But the child doesnt exist in any environment, and
its unclear if the child term should be created OR if we should change the logic to fallback to parent if one exists but child doesnt.

In this case child:
                    "element_id": 30,
                    "code": "SH20_121",
                    "start_date": "2019-07-31T21:00:00.000Z",
                    "end_date": "2026-07-30T21:00:00.000Z",
                    "name": [
                        {
                            "langcode": "fi",
                            "text": "Oikeustieteen koulutus, Helsinki"
                        },

Parent:
                    "element_id": 20,
                    "code": "KH20_001",
                    "start_date": "2019-07-31T21:00:00.000Z",
                    "end_date": "2026-07-30T21:00:00.000Z",
                    "name": [
                        {
                            "langcode": "fi",
                            "text": "Oikeusnotaarin koulutusohjelma"
..

The combination would be "[code] => KH20_001SH20_121" - possibly something like:
"Oikeusnotaarin koulutusohjelma - oikeustieteen koulutus, Helsinki"

But we dont have such an degree_programme term in production or anywhere. So "Oikeusnotaarin koulutusohjelma" is actually automatically flagged for the person but its "is_primary" field of the personal flagging remains FALSE, so its visible in the list under "My degree programmes" but its not automatically activated.