HUB-777 - Enable bigpipe (core) module to improve perceived performance for dynamic/user content!
Enabling the module does not require further Drupal config changes (and apparently also no customizations).

For server side, we need to alter varnish config slightly to avoid buffering issues:
https://www.drupal.org/docs/8/core/modules/big-pipe/bigpipe-environment-requirements

  # Disable buffering only for BigPipe responses
  if (beresp.http.Surrogate-Control ~ "BigPipe/1.0") {
    set beresp.do_stream = true;
    set beresp.ttl = 0s;
  }

(Note: we might not notice a difference locally but in QA/PROD as these changed were implemented manually for those envs (documented in separate infra PR, but docker is not handled there as I didnt find an easy way to do it).